Refaz a estrutura dos arquivos de script para apenas 1
