+++ index.html
@@ -4,19 +4,12 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <script defer src="/src/scripts/pokedex.js"></script>
+  <script defer src="script.js"></script>
   <title>Who's that Pokemon</title>
 </head>
 <body>
   <img src="pokemon.png" alt="" id="pokemon-image">
-  <section id="alternatives-field">
-    <input type="radio" id="pokemon1" name="fav_language" value="pikachu">
-    <label for="pokemon1">pokemon</label><br>
-    <input type="radio" id="pokemon2" name="fav_language" value="squirdle">
-    <label for="pokemon2">pokemon</label><br>
-    <input type="radio" id="pokemon3" name="fav_language" value="meowth">
-    <label for="pokemon3">pokemon</label>
-  </section>
+  <section id="alternatives-field"></section>
   <button id="confirm">Confirm</button>
 </body>
 </html>
