--- src/scripts/api.js
@@ -1,11 +1,0 @@
-const fetch = require('node-fetch');
-
-async function getPokemonById(id) {
-  const response = await fetch(`https://pokeapi.co/api/v2/pokemon/${id}/`);
-  const pokeObject = await response.json();
-  return pokeObject;
-}
-
-module.exports = {
-  getPokemonById,
-}
