--- src/scripts/pokedex.js
@@ -1,10 +1,0 @@
-const { getPokemonById } = require('./api');
-const { randomNumber } = require('./utils');
-
-async function getRandomPokemon() {
-  const number = randomNumber(1, 898);
-  const pokemon = await getPokemonById(number);
-  return pokemon;
-}
-
-
