--- src/scripts/utils.js
@@ -1,9 +1,0 @@
-function getRandomIntInclusive(min, max) {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-module.exports = {
-  randomNumber: getRandomIntInclusive,
-}
